backend/requirements.txt
# Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1
pydantic>=2.5.3
pydantic-settings>=2.1.0

# PDF Processing
pymupdf>=1.23.8
pytesseract>=0.3.10
Pillow>=10.2.0

# AI & Embeddings
openai>=1.10.0
tiktoken>=0.5.2

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
aiofiles>=23.2.1
pgvector>=0.2.4
greenlet
google-auth==2.45.0
passlib==1.7.4
python-jose==3.5.0
python-multipart==0.0.20
supabase
bcrypt>=4.0.0
